add workforceHelpers tests and move the appropriate workforce tests
